fixed all warnings and fixed clangd issues

+++ llarm-emu/src/cpu/instructions/arm/logic.cpp
@@ -52,7 +52,7 @@ void INSTRUCTIONS::arm::logic::CLZ(const u32 code) {
     if (Rm == 0) {
         reg.write(Rd_id, 32);
     } else {
-        reg.write(Rd_id, 31 - (std::bit_width(Rm) - 1));
+        reg.write(Rd_id, 31 - u32(std::bit_width(Rm) - 1));
     }
 }
