+++ llarm-emu/src/cpu/memory/fcse.cpp
@@ -16,8 +16,8 @@ u32 FCSE::modify_address(const u32 address) {
         return address;
     }
 
-    const u8 PID = coprocessor.read(id::cp15::R13_PID);
-    return (address | PID << 25);
+    const u8 PID = static_cast<u8>(coprocessor.read(id::cp15::R13_PID));
+    return (address | u32(PID << 25));
 }
 
 // TODO ctrl+f for "FCSE" in the arm docs once all the MMU, cache, and write buffer stuff are complete
